Delay saving new users as long as possible.
LDAPBackend.get_or_create() no longer saves a newly created user but instead
simply instantiates it. This allows NOT NULL fields to be populated by
USER_ATTR_MAP or the populate_user signal and avoids an IntegrityError.

To test, added a new NOT NULL field to TestUser. This is populated by
USER_ATTR_MAP in the tests.

Fixes #58